Show deprecation warning while toggling bottomTabs visibility on iOS (#6409)
On iOS the system doesn't support dynamically toggling bottomTabs at all so we had to hack through it which resulted in many issues with this feature. We plan on entirely removing it and stick to the framework's guidelines which means that toggling the bottomTabs will not be possible at all and controlling it with static options will be the only option.