Another Shared Element Demo for Playground (#6516)
I've added another use-case for Shared Element Transitions to the Playground app. The main differences to the cocktail one are;

* Full-Screen Details Screen (no statusbar, no titlebar, no bottom tabs)
* Spring animation (added prototype iOS native code for now, android not yet implemented!)
* More physical feeling through scale button press feedback, spring animation, etc

Currently there are some issues with the SET implementation, which are nicely visible in this demo.

1. Images don't really work (different aspect ratios) and get glitchy
2. There are problems with the StatusBar and TitleBar
3. There are problems with the bottom tabs animation

All of these are known issues and have separate GitHub issues created.

Also, I wanted to add a drag down animation to dismiss the details screen using gesture handler and reanimated, but I couldn't get the libraries installed. Will look into this.

Co-authored-by: Jin Shin <jinshin1013@gmail.com>
Co-authored-by: mrousavy <mrousavy@donotreply.github.com>
Co-authored-by: Guy Carmeli <guyca@users.noreply.github.com>
Co-authored-by: yogevbd <yogev132@gmail.com>