Remove unnecessary overrides to setDefaultOptions, apply defaultOptions to RNNSideMenuController (#6691)
* Removes `setDefaultOptions` from RNNBasePresenter. This is already covered by the synthesized `defaultOptions` prop.
* Add calling `[presenter setDefaultOptions:]` to the default implementation in order to remove all the overrides that do the same thing.

Notice that the override implementation of `setDefaultOptions` in `RNNSideMenuController` does not call `super`. This leads to the `defaultOptions` only being available in the `RNNSideMenuPresenter` but not in the controller itself. Therefor causing bugs in accessing options that were set as defaults through the controller (as they are just not available).